backend/requirements.txt
pymongo==4.3.3
python-dotenv==0.21.0
pyjwt==2.6.0
pytest==7.2.0
pytest-mock==3.10.0
boto3==1.26.27
botocore==1.29.27
mangum==0.17.0
pydantic==1.10.2
fastapi==0.89.1
uvicorn==0.20.0 